requirements.txt
Flask==2.0.1
PyMySQL==1.0.2
Werkzeug==2.0.1
cryptography==3.4.7
requests==2.31.0
urllib3==1.26.20
chardet==5.2.0
charset-normalizer==2.0.12
python-dotenv==0.19.0 